@@ -3,21 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
     "axios": "^0.19.0",
+    "axios-mock-adapter": "^1.17.0",
+    "dateformat": "^3.0.3",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "jest": "^24.9.0",
+    "jest-environment-jsdom-fourteen": "^0.1.0",
+    "jquery": "^3.4.1",
     "materialize-css": "^1.0.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
+    "query-string": "^6.11.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-helmet": "^5.2.1",
-    "react-materialize": "^3.4.1",
+    "react-image": "^2.4.0",
+    "react-redux": "^7.1.3",
+    "react-responsive-carousel": "^3.1.51",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0"
+    "react-scripts": "3.2.0",
+    "react-side-effect": "^2.1.0",
+    "redux": "^4.0.4",
+    "sanitize-html": "^1.20.1",
+    "sinon": "^7.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "test": "react-scripts test --env=jsdom-fourteen",
+    "eject": "react-scripts eject",
+    "coverage": "react-scripts test --env=jsdom-fourteen --watchAll=false --coverage"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -32,6 +48,13 @@
       "last 1 chrome version",
       "last 1 firefox version",
       "last 1 safari version"
+    ]
+  },
+  "jest": {
+    "coveragePathIgnorePatterns": [
+      "/node_modules/",
+      "/src/index.js",
+      "/src/serviceWorker.js"
     ]
   }
 }
