@@ -44,15 +44,15 @@ class Navbar extends Component {
           <NavItem label="The Team" to="/about/team" />
           <NavItem label="Contact Us" to="/about/contact" />
         </ul>
-        <nav>
+        <nav id="main-navigation">
           <div className="nav-wrapper">
             <div className="container">
               <a href="#!" className="brand-logo left">{config.appName}</a>
               <ul id="nav-mobile" className="right">
                 <NavItem label="Home" to="/" />
                 <NavItem label="Features" to="/features" />
-                <NavItem label="Help" to="/help" />
                 <NavItem label="Get It" to="/product" />
+                <NavItem label="Help" to="/help" />
                 <li className={isNested ? "active" : ""}>
                   <a className="dropdown-trigger" href="#!"
                      data-target="about-dropdown">About</a>
