@@ -4,8 +4,10 @@ import Layout from './Layout';
 
 const NotFound = () => (
   <Layout pageTitle="Not Found">
-    <p className="text-center"><h4>404 - Not Found</h4></p>
-    <p className="text-center">Visit <Link to="/">Home</Link> to find what you're looking for.</p>
+    <h4 className="textCenter">404 - Not Found</h4>
+    <p className="textCenter">
+      Visit <Link to="/">Home</Link> to find what you're looking for.
+    </p>
   </Layout>
 );
 
